PrivateRoute should also check for refresh token expiration
This prevents a situaion when React Router redirects to some component,
this component tries to call the API and only then the user gets
redirected to the login page. The problem here is that the user can see
the new component for a split second - this looks bad. So now
PrivateRoute also checks is the user should be totally logged out and if
so.